+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## Unreleased changes
 
+## 0.1.2.0
+
+- [#6](https://github.com/parsonsmatt/annotated-exception/pull/6)
+    - Add `Control.Exception.Annotated.UnliftIO` that uses `MonadUnliftIO`
+      instead of `MonadCatch` and `MonadThrow`.
+    - Actually expose `catches`
+
 ## 0.1.1.0
 
 - [#4](https://github.com/parsonsmatt/annotated-exception/pull/4)